requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
django-cors-headers==4.3.1
